# services/api/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Anbaro API",
    "version": "1.0.0",
    "description": "Generated from route Zod schemas."
  },
  "paths": {
...
    "/api/v1/assistant/stock-proposals": {
      "post": {
        "summary": "Turn a natural-language message into a stock-change proposal (no write)",
        "tags": [
          "assistant"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "locationId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  }
                },
                "required": [
                  "message"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "2XX": {
            "description": "Success"
          },
          "4XX": {
            "description": "Client error"
          }
        }
      }
    },
    "/api/v1/assistant/interactions": {
      "post": {
        "summary": "Record what the user did with an assistant proposal",
        "tags": [
          "assistant"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transcriptId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "message": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "outcomes": {
                    "minItems": 1,
                    "maxItems": 25,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "outcome": {
                          "type": "string",
                          "enum": [
                            "confirmed",
                            "corrected",
                            "rejected"
                          ]
                        },
                        "proposed": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "corrected": {
                          "nullable": true,
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "outcome",
                        "proposed"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "required": [
                  "transcriptId",
                  "message",
                  "outcomes"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "2XX": {
            "description": "Success"
          },
          "4XX": {
            "description": "Client error"
          }
        }
      }
    }
  }
}